+++ internal/messenger/selectors.go
@@ -1,0 +1,10 @@
+package messenger
+
+const (
+	feedCardSelector      = "div.a11y_feed_card_item"
+	feedTitleSelector     = "div._2b17aec6"
+	recentScrollSelector  = "div.simplebar-content-wrapper"
+	searchTriggerSelector = ".appNavbar-search-input-wrapper"
+	chatTitleSelector     = "#root-messenger-chat span.chatWindow_chatName"
+	rightPanelSelector    = "#root-messenger-chat"
+)